ported board method and state to Game
