+++ src/Game.ts
@@ -1,0 +1,29 @@
+import { Piece, Square, PieceSymbol, Color, Ox88, algebraic } from './types'
+
+export class Game {
+  _board = new Array<Piece>(128)
+
+  board(): ({ square: Square; type: PieceSymbol; color: Color } | null)[][] {
+    const output = []
+    let row = []
+
+    for (let i = Ox88.a8; i <= Ox88.h1; i++) {
+      if (this._board[i] == null) {
+        row.push(null)
+      } else {
+        row.push({
+          square: algebraic(i),
+          type: this._board[i].type,
+          color: this._board[i].color,
+        })
+      }
+      if ((i + 1) & 0x88) {
+        output.push(row)
+        row = []
+        i += 8
+      }
+    }
+
+    return output
+  }
+}
