+++ src/chess.ts
@@ -25,6 +25,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
+import { Game } from './Game'
 import { Move } from './Move'
 import { parse } from './pgn'
 import {
@@ -45,6 +46,7 @@ import {
   rank,
   file,
   algebraic,
+  Ox88,
 } from './types'
 
 // Re-export types and constants from types.ts for backward compatibility
@@ -233,19 +235,6 @@ const HEADER_TEMPLATE = {
  *    array comes in. It provides an offset (in this case 16) to add to E7 (20)
  *    to check for blocking pieces. E7 (20) + 16 = E6 (36) + 16 = E5 (52) etc.
  */
-
-// prettier-ignore
-// eslint-disable-next-line
-const Ox88: Record<Square, number> = {
-  a8:   0, b8:   1, c8:   2, d8:   3, e8:   4, f8:   5, g8:   6, h8:   7,
-  a7:  16, b7:  17, c7:  18, d7:  19, e7:  20, f7:  21, g7:  22, h7:  23,
-  a6:  32, b6:  33, c6:  34, d6:  35, e6:  36, f6:  37, g6:  38, h6:  39,
-  a5:  48, b5:  49, c5:  50, d5:  51, e5:  52, f5:  53, g5:  54, h5:  55,
-  a4:  64, b4:  65, c4:  66, d4:  67, e4:  68, f4:  69, g4:  70, h4:  71,
-  a3:  80, b3:  81, c3:  82, d3:  83, e3:  84, f3:  85, g3:  86, h3:  87,
-  a2:  96, b2:  97, c2:  98, d2:  99, e2: 100, f2: 101, g2: 102, h2: 103,
-  a1: 112, b1: 113, c1: 114, d1: 115, e1: 116, f1: 117, g1: 118, h1: 119
-}
 
 const PAWN_OFFSETS = {
   b: [16, 32, 17, 15],
@@ -580,7 +569,7 @@ function strippedSan(move: string): string {
 }
 
 export class Chess {
-  private _board = new Array<Piece>(128)
+  private _game!: Game
   private _turn: Color = WHITE
   private _header: Record<string, string | null> = {}
   private _kings: Record<Color, number> = { w: EMPTY, b: EMPTY }
@@ -599,13 +588,18 @@ export class Chess {
   private _positionCount = new Map<bigint, number>()
 
   constructor(fen = DEFAULT_POSITION, { skipValidation = false } = {}) {
+    this._game = new Game()
     this._comments = {}
     this._suffixes = {}
     this.load(fen, { skipValidation })
   }
 
+  private get _board() {
+    return this._game._board
+  }
+
   clear({ preserveHeaders = false } = {}) {
-    this._board = new Array<Piece>(128)
+    this._game._board = new Array<Piece>(128)
     this._kings = { w: EMPTY, b: EMPTY }
     this._turn = WHITE
     this._castling = { w: 0, b: 0 }
@@ -2414,27 +2408,7 @@ export class Chess {
   }
 
   board(): ({ square: Square; type: PieceSymbol; color: Color } | null)[][] {
-    const output = []
-    let row = []
-
-    for (let i = Ox88.a8; i <= Ox88.h1; i++) {
-      if (this._board[i] == null) {
-        row.push(null)
-      } else {
-        row.push({
-          square: algebraic(i),
-          type: this._board[i].type,
-          color: this._board[i].color,
-        })
-      }
-      if ((i + 1) & 0x88) {
-        output.push(row)
-        row = []
-        i += 8
-      }
-    }
-
-    return output
+    return this._game.board()
   }
 
   squareColor(square: Square): 'light' | 'dark' | null {
